// src/shared/api/generated/model/updateTeamMemberRequestDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTeamMemberRequestDtoPermissionsItem } from './updateTeamMemberRequestDtoPermissionsItem.ts';
import type { UpdateTeamMemberRequestDtoRole } from './updateTeamMemberRequestDtoRole.ts';
import type { UpdateTeamMemberRequestDtoStatus } from './updateTeamMemberRequestDtoStatus.ts';

export interface UpdateTeamMemberRequestDto {
  role?: UpdateTeamMemberRequestDtoRole;
  status?: UpdateTeamMemberRequestDtoStatus;
  /** @maxItems 6 */
  permissions?: UpdateTeamMemberRequestDtoPermissionsItem[];
  /**
     * @minLength 1
     * @maxLength 80
     * @nullable
     */
  staffTitle?: string | null;
}